requirements.txt
asgiref==3.8.1
cffi==1.17.1
Django==5.0
django-cors-headers==4.5.0
django-jazzmin==3.0.0
djangorestframework==3.15.2
gunicorn==23.0.0
h3==3.7.7
islam==2.2.0
mysqlclient==2.2.5
numpy==2.1.1
packaging==24.1
pillow==10.4.0
pycparser==2.22
pytz==2024.2
setuptools==75.1.0
sqlparse==0.5.1
timezonefinder==6.5.3
